server/requirements.txt
flask
python-socketio
influxdb-client
flask-cors